==> Datasets Tools/step2-stardist-nuclei-segmentation.ipynb <==
#Coppied and modified from Tamas Nagy code 2021: 
#Evelyn Strickland 2022, for use on my imaging setup, multithreading doesn't work just yet with this????
#V1.0, 2024, included in toolset for analysis pipeline!

from csbdeep.utils import normalize
from tqdm.notebook import tqdm as tqdm
from pathlib import Path
from stardist.models import StarDist2D 
from tifffile import TiffFile, TiffWriter
import tensorflow as tf

#General command to run stardist on a 2 channel tif, or a single channel nuclei only tif: 
def run_stardist(filepath, model):
    
    #Get the filename directory, ome-tag, create saveloc, and virutally open the image: 
    dirname = filepath.parent
    omename = filepath.name.split('.', 1)[0]
    savedir = filepath.parent.parent / 'segmented_nuclei'
    img = TiffFile(filepath)
    
    #Open a writer to output each slice to a new segmeneted tif at saveloc: 
    with TiffWriter(savedir.joinpath(omename + "_segmented.tif")) as tif:
        page = img.pages
        series = img.series[0]
        #Check to see if the images are both a calcium and nuclei channel. If so, we expect the images to have the calcium channel FIRST, nuclei SECOND
        if len(page) == 2 * series.get_shape()[0]:
            pages = page[1::2]
        else:
            pages = page
        #Take the pages, normalize them, then pass them over to the stardist model for segmentation: 
        for page in tqdm(pages):
            pg = normalize(page.asarray(), 1, 99.8, axis=(0,1))
            labels, _ = model.predict_instances(pg)
            tif.write(labels.astype('uint16'), contiguous=True)


#import the model from 20220902 custom trained model (included in archive)
model = StarDist2D(None, name='stardist', basedir='20220902_model')


#SET THIS BEOFORE RUNNING:
folder = '' #Dataset you want to analyze.
folderpath = Path(folder, )
img_folder = folderpath / 'images'
img_paths = list(img_folder.glob('*.tif'))

#Quick dummy check: 
if len(img_paths) == 0:
    print("Check filepath, no images detected...")


#Main Running Loop: 
for file in img_paths:
    
    dirname = file.parent
    omename = file.name.split('.', 1)[0]
    savedir = file.parent.parent / 'segmented_nuclei'
    
    #If the segmentation file already exists, skip this in the loop
    if savedir.joinpath(omename + "_segmented.tif").exists():
        print('Image {} already has a segmentation!'.format(omename))
        continue
    
    print('Running stardist on {}.tif'.format(file.stem))
    run_stardist(file,model)
